+++ docs/changelog.rst
@@ -11,6 +11,8 @@ Unreleased
 
 Added
 -----
+- New generic methods for ``GridModel``: ``setup_grid``, ``setup_grid_from_constant``, ``setup_grid_from_raster``, ``setup_grid_from_raster_reclass``, ``setup_grid_from_vector``.
+- New raster method ``rasterize_geometry``.
 
 Changed
 -------